A very naive non-responsive website for easy learning and grasp core concepts of HTML4 and CSS.
There are two HTML files and a CSS file. It is basically a part of Homework assignment of CSCI 571 Web Development course. Whole purpose for was learning core concepts of HTML and CSS from basic to advanced. There are links home and nature which redirect from one page to another. All other links are clickable but no redirect action is applied.